Added 'void Idle(void)'.  proc.c is complete.
The function Idle() accepts nothing and returns nothing. It runs an Infinite while loop which flickers the dot after every second. incrementing the sys_time_count inside the while loop, when it hits a multiple of 100 we check for the flag. depending on it we show or hide the bold faced dot.